New component for statustracking, fix profile and user_settings
