+++ src/resources/js/Components/InactivityTracker.jsx
@@ -1,5 +1,4 @@
-import { useCallback, useContext, useEffect, useRef } from "react";
-import { usePage } from '@inertiajs/react';
+import { useCallback, useContext, useEffect } from "react";
 import { ApplicationContext } from "@/Components/ApplicationContext.jsx";
 
 export default function InactivityTracker() {
@@ -7,14 +6,9 @@ export default function InactivityTracker() {
         setUserPublicKey,
         setUserPrivateKey,
         isInactive, setIsInactive,
-        sessionLocked, pageIsHidden,
     } = useContext(ApplicationContext);
 
-    const userId = usePage().props.auth.user.id;
-
     const inactivityTimeout = (import.meta.env.VITE_INACTIVITY_TIMEOUT ?? 600) * 1000;
-
-    const isConnectedRef = useRef(false);
 
     const resetTimer = useCallback(() => {
         setIsInactive(false);
@@ -39,21 +33,11 @@ export default function InactivityTracker() {
     }, [resetTimer]);
 
     useEffect(() => {
-        const shouldBeOnline = !isInactive && !sessionLocked && !pageIsHidden;
-
-        if (shouldBeOnline && !isConnectedRef.current) {
-            Echo.private(`user-status.${userId}`);
-            isConnectedRef.current = true;
-        } else if (!shouldBeOnline && isConnectedRef.current) {
-            Echo.leave(`user-status.${userId}`);
-            isConnectedRef.current = false;
-        }
-
         if (isInactive) {
             sessionStorage.clear();
 
             setUserPublicKey(null);
             setUserPrivateKey(null);
         }
-    }, [isInactive, sessionLocked, pageIsHidden]);
+    }, [isInactive]);
 }
